refactor(architecture): move transcribeAudio to shared audioService
Extract transcribeAudio from Journey's momentPersistenceService to
src/services/audioService.ts, fixing a DDD bounded context violation.

This function was used by 7 files across 4 modules (hooks, components,
services, journey). Moving it to shared services respects module boundaries.

Closes #730
